Move functions into base_tasks.py
