+++ Lib/asyncio/base_tasks.py
@@ -3,6 +3,7 @@ import traceback
 
 from . import base_futures
 from . import coroutines
+from .events import get_running_loop
 
 
 def _task_repr_info(task):
@@ -74,3 +75,63 @@ def _task_print_stack(task, limit, file):
     if exc is not None:
         for line in traceback.format_exception_only(exc.__class__, exc):
             print(line, file=file, end='')
+
+
+# dict of {EventLoop: Set[tasks]} containing all tasks alive.
+_all_tasks = {}
+
+# Dictionary containing tasks that are currently active in
+# all running event loops.  {EventLoop: Task}
+_current_tasks = {}
+
+
+def current_task(loop=None):
+    if loop is None:
+        loop = get_running_loop()
+    task = _current_tasks.get(loop)
+    if task is None:
+        raise RuntimeError("no current task")
+    return task
+
+
+def all_tasks(loop=None):
+    if loop is None:
+        loop = get_running_loop()
+    task = _current_tasks.get(loop)
+    if task is None:
+        raise RuntimeError("no current task")
+    return task
+
+
+def _register_task(loop, task):
+    tasks = _all_tasks.get(loop)
+    if tasks is None:
+        tasks = _all_tasks[loop] = set()
+    tasks.add(task)
+
+
+def _enter_task(loop, task):
+    current_task = _current_tasks.get(loop)
+    if current_task is not None:
+        raise RuntimeError(f"Entering into task {task!r} "
+                           f"when other task {current_task!r} is executed.")
+    _current_tasks[loop] = task
+
+
+def _leave_task(loop, task):
+    current_task = _current_tasks.get(loop)
+    if current_task is not task:
+        raise RuntimeError(f"Leaving task {task!r} "
+                           f"is not current {current_task!r}.")
+    del _current_tasks[loop]
+
+
+def _unregister_task(loop, task):
+    tasks = _all_tasks.get(loop)
+    if tasks is None:
+        # loop is not registered
+        # raise RuntimeWarning?
+        return
+    tasks.remove(task)
+    if not tasks:
+        del _all_tasks[loop]
